Ajustes en el modulo de Listar y Cargar interesados
